move getKey to useCharactersApi
